I changed the home page list to look a bit more visually appealing.
This was for the block and inline lesson, and what I did was
create classes for the list items, make them inline boxes and
changed their borders and background colors, as with their
height and widths to make it look more visually appealing.
I am happy with my decisions although it would be difficult
to grow and add more recipies easily with this design.